fix sign out error
