--- src/components/navbar/auth/auth-options.tsx
@@ -1,80 +1,0 @@
-"use client";
-
-import { ReactNode, useTransition } from "react";
-import { logout } from "./actions";
-import { Loader2 } from "lucide-react";
-import { TokenPayload } from "@/lib/universal/auth";
-import UserDropdown from "./user-dropdown";
-import SignInDialog from "./sign-in/sign-in-dialog";
-import { CosmoArtistBFF } from "@/lib/universal/cosmo/artists";
-import { ValidArtist } from "@/lib/universal/cosmo/common";
-import { PublicProfile } from "@/lib/universal/cosmo/auth";
-import GuestThemeSwitch from "./guest-theme-switch";
-import Hydrated from "@/components/hydrated";
-import { useWallet } from "@/hooks/use-wallet";
-import { usePathname } from "next/navigation";
-// import GuestArtistSwitch from "./guest-artist-switch";
-
-type Props = {
-  token: TokenPayload | undefined;
-  profile: PublicProfile | undefined;
-  artists: CosmoArtistBFF[];
-  selectedArtist: ValidArtist;
-  comoBalances: ReactNode;
-  cosmoAvatar: ReactNode;
-};
-
-export default function AuthOptions({
-  token,
-  profile,
-  artists,
-  selectedArtist,
-  comoBalances,
-  cosmoAvatar,
-}: Props) {
-  const pathname = usePathname();
-  const [isPending, startTransition] = useTransition();
-  const { disconnect } = useWallet();
-
-  const isAuthenticated = !!token && !!profile;
-
-  function executeSignOut() {
-    startTransition(async () => {
-      disconnect();
-      await logout();
-    });
-  }
-
-  if (isAuthenticated) {
-    return (
-      <div className="flex gap-2 items-center justify-center">
-        <div className="md:flex gap-2 items-center hidden">{comoBalances}</div>
-
-        <UserDropdown
-          key={pathname}
-          profile={profile}
-          artists={artists}
-          selectedArtist={selectedArtist}
-          comoBalances={comoBalances}
-          onSignOut={executeSignOut}
-          cosmoAvatar={cosmoAvatar}
-        />
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex items-center gap-2">
-      {isPending ? <Loader2 className="animate-spin" /> : <SignInDialog />}
-
-      <Hydrated>
-        <GuestThemeSwitch />
-      </Hydrated>
-
-      {/* <GuestArtistSwitch
-      artists={artists}
-      selectedArtist={selectedArtist}
-    /> */}
-    </div>
-  );
-}
