+++ src/components/navbar/auth/sign-in/sign-in-dialog.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import {
   AlertDialog,
   AlertDialogContent,
